+++ src/events/banlist.h
@@ -1,0 +1,7 @@
+#ifndef GUARD_BANLIST_H
+#define GUARD_BANLIST_H
+
+void event_banlist(struct irc_message_compo *);
+void event_eof_banlist(struct irc_message_compo *);
+
+#endif